# training_code/module5_adv_boot_ms/code_adv_boot/productappv6topdown/src/main/resources/swagger/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8090
  description: Generated server url
paths:
  /productapp/v1/products/{id}:
    put:
      tags:
      - product-controller-v-1
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDto'
    delete:
      tags:
      - product-controller-v-1
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: No Content
  /productapp/v1/products:
    get:
      tags:
      - product-controller-v-1
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
    post:
      tags:
      - product-controller-v-1
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: string
  /productapp/v1/productsbyname/{name}:
    get:
      tags:
      - product-controller-v-1
      operationId: getByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
  /productapp/v1/products/v1/{id}:
    get:
      tags:
      - product-controller-v-1
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDto'
  /infodto:
    get:
      tags:
      - info-dto-controller
      operationId: infoDto
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InfoDto'
components:
  schemas:
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
    ProductDto:
      required:
      - name
      - price
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          pattern: "[A-Za-z]+( [A-Za-z]+)*"
          type: string
        price:
          type: number
        _links:
          $ref: '#/components/schemas/Links'
    InfoDto:
      required:
      - message
      type: object
      properties:
        message:
          type: string
        contactDetails:
          type: object
          additionalProperties:
            type: string
        onCallSupport:
          type: array
          items:
            type: string
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean