# training_code/module5_adv_boot_ms/code_adv_boot/productappv7caching/src/main/java/com/productapp/dto/InfoDto.java
package com.productapp.dto;

import jakarta.validation.constraints.NotEmpty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.ToString;
import org.springframework.boot.context.properties.ConfigurationProperties;

import java.util.List;
import java.util.Map;
@Data
@NoArgsConstructor
@AllArgsConstructor
@ToString
@ConfigurationProperties(prefix = "account")
public class InfoDto {

    @NotEmpty
    private String message;
    private Map<String ,String> contactDetails;
    private List<String> onCallSupport;
}